Moves EntitlementInfo
